sprint1/step1: добавил тег <a> и удалил дубликаты propTypes
