+++ src/components/AppHeader/AppHeader.module.css
@@ -53,3 +53,8 @@
     align-items: center;
 }
 
+a {
+    text-decoration: none;
+    color: var(--text-inactive-color) !important;
+}
+
