+++ src/components/BurgerIngredients/IngredientsType/IngredientsType.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import styles from './IngredientsType.module.css'
 import IngredientCard from "./IngredientCard/IngredientCard";
 import PropTypes from "prop-types";
+import DataPropTypes from '../../../utils/prop-types';
 
 function IngredientsType(props) {
     const dataObj = {
@@ -30,21 +31,6 @@ function IngredientsType(props) {
         </>
     );
 }
-
-const DataPropTypes = PropTypes.shape({
-    calories: PropTypes.number.isRequired,
-    carbohydrates: PropTypes.number.isRequired,
-    fat: PropTypes.number.isRequired,
-    image: PropTypes.string.isRequired,
-    image_large: PropTypes.string.isRequired,
-    image_mobile: PropTypes.string.isRequired,
-    name: PropTypes.string.isRequired,
-    price: PropTypes.number.isRequired,
-    proteins: PropTypes.number.isRequired,
-    type: PropTypes.oneOf(['main', 'bun', 'sauce']),
-    __v: 0,
-    _id: PropTypes.string.isRequired,
-});
 
 IngredientsType.propTypes = {
     data: PropTypes.arrayOf(DataPropTypes).isRequired
